src/requirements.txt
Flask
python-dotenv
PyTelegramBotAPI
numpy
Pillow